# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Roblox Vector Search API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/commands/gather-games": {
      "post": {
        "summary": "Start a job to gather games from Roblox API",
        "operationId": "gatherGames",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Job started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "jobId": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "pending",
                            "running",
                            "completed",
                            "failed"
                          ]
                        }
                      },
                      "required": [
                        "jobId",
                        "message",
                        "status"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "success": true,
                  "data": {
                    "jobId": "string",
                    "message": "string",
                    "status": "string"
                  }
                }
              }
            }
          },
          "default": {
            "description": "Global error response",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Error message"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "data": false
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}